refactor: ProductDetailPage 로직을 Custom hook으로 분리
